sys/include/libc.h: use uintptr in offsetof (thanks Geoff Collyer)

+++ sys/include/libc.h
@@ -2,7 +2,7 @@
 #pragma	src	"/sys/src/libc"
 
 #define	nelem(x)	(sizeof(x)/sizeof((x)[0]))
-#define	offsetof(s, m)	(ulong)(&(((s*)0)->m))
+#define	offsetof(s, m)	(uintptr)(&(((s*)0)->m))
 #define	assert(x)	if(x){}else _assert("x")
 
 /*
